+++ src/main/java/gift/config/KakaoProperties.java
@@ -1,4 +1,4 @@
-package gift.auth;
+package gift.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
